Permission set in controllers, repositories, services and views

+++ app/Http/Controllers/MenuController.php
@@ -17,29 +17,40 @@ class MenuController extends BaseController
 
     public function index()
     {
-        $this->setPageData('Menu','Menu','fas fa-th-list');
-        return view('menu.index');
+        if(permission('menu-access')){
+            $this->setPageData('Menu','Menu','fas fa-th-list');
+            return view('menu.index');
+        }else{
+            return $this->unauthorized_access_blocked();
+        }
+       
     }
 
     public function get_datatable_data(Request $request)
     {
-        if($request->ajax()){
-            $output = $this->service->get_datatable_data($request);
-        }else{
-            $output = ['status'=>'error','message'=>'Unauthorized Access Blocked!'];
-        }
+        if(permission('menu-access')){
+            if($request->ajax()){
+                $output = $this->service->get_datatable_data($request);
+            }else{
+                $output = ['status'=>'error','message'=>'Unauthorized Access Blocked!'];
+            }
 
-        return response()->json($output);
+            return response()->json($output);
+        }
     }
 
     public function store_or_update_data(MenuRequest $request)
     {
         if($request->ajax()){
-            $result = $this->service->store_or_update_data($request);
-            if($result){
-                return $this->response_json($status='success',$message='Data Has Been Saved Successfully',$data=null,$response_code=200);
+            if(permission('menu-add') || permission('menu-edit')){
+                $result = $this->service->store_or_update_data($request);
+                if($result){
+                    return $this->response_json($status='success',$message='Data Has Been Saved Successfully',$data=null,$response_code=200);
+                }else{
+                    return $this->response_json($status='error',$message='Data Cannot Save',$data=null,$response_code=204);
+                }
             }else{
-                return $this->response_json($status='error',$message='Data Cannot Save',$data=null,$response_code=204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
@@ -49,11 +60,15 @@ class MenuController extends BaseController
     public function edit(Request $request)
     {
         if($request->ajax()){
-            $data = $this->service->edit($request);
-            if($data->count()){
-                return $this->response_json($status='success',$message=null,$data=$data,$response_code=201);
+            if(permission('menu-edit')){
+                $data = $this->service->edit($request);
+                if($data->count()){
+                    return $this->response_json($status='success',$message=null,$data=$data,$response_code=201);
+                }else{
+                    return $this->response_json($status='error',$message='No Data Found',$data=null,$response_code=204);
+                }
             }else{
-                return $this->response_json($status='error',$message='No Data Found',$data=null,$response_code=204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
@@ -63,11 +78,15 @@ class MenuController extends BaseController
     public function delete(Request $request)
     {
         if($request->ajax()){
-            $result = $this->service->delete($request);
-            if($result){
-                return $this->response_json($status='success',$message='Data Has Been Deleted Successfully',$data=null,$response_code=200);
+            if(permission('menu-delete')){
+                $result = $this->service->delete($request);
+                if($result){
+                    return $this->response_json($status='success',$message='Data Has Been Deleted Successfully',$data=null,$response_code=200);
+                }else{
+                    return $this->response_json($status='error',$message='Data Cannot Delete',$data=null,$response_code=204);
+                }
             }else{
-                return $this->response_json($status='error',$message='Data Cannot Delete',$data=null,$response_code=204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
@@ -77,11 +96,15 @@ class MenuController extends BaseController
     public function bulk_delete(Request $request)
     {
         if($request->ajax()){
-            $result = $this->service->bulk_delete($request);
-            if($result){
-                return $this->response_json($status='success',$message='Data Has Been Deleted Successfully',$data=null,$response_code=200);
+            if(permission('menu-bulk-delete')){
+                $result = $this->service->bulk_delete($request);
+                if($result){
+                    return $this->response_json($status='success',$message='Data Has Been Deleted Successfully',$data=null,$response_code=200);
+                }else{
+                    return $this->response_json($status='error',$message='Data Cannot Delete',$data=null,$response_code=204);
+                }
             }else{
-                return $this->response_json($status='error',$message='Data Cannot Delete',$data=null,$response_code=204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
